% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_clust.R
\name{em_clust}
\alias{em_clust}
\title{Perform the Expectation-Maximization Algorithm on a dataset}
\usage{
em_clust(dat, k)
}
\arguments{
\item{dat}{a numeric dataset to perform the algorithm on}

\item{k}{The number of random observations to start with}
}
\value{
A list with cluster assignments and parameter estimates
}
\description{
Perform the Expectation-Maximization Algorithm on a dataset
}
\examples{

}
